requirements.txt
# Core libraries
numpy
pandas
scikit-learn
joblib
PyYAML
spacy
matplotlib
seaborn


# NLP and ML
tensorflow
tf-keras
torch
transformers
datasets
networkx
py2neo
nltk

# Web scraping & APIs
requests
beautifulsoup4
tldextract
fastapi
uvicorn
